requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.32.2
selenium==4.22.0
tqdm==4.66.4
black==24.4.2
flake8==7.1.0
mypy-extensions==1.0.0
isort==5.13.2